Rewrote the part types UI to fix it, it was poorly implemented. This is a nested structure that will have much better performance and should be easier to use

+++ Binner/Library/Binner.Common/Services/PartTypeService.cs
@@ -1,5 +1,7 @@
 ﻿using Binner.Model.Common;
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Binner.Common.Services
@@ -22,6 +24,17 @@ namespace Binner.Common.Services
 
         public async Task<bool> DeletePartTypeAsync(PartType partType)
         {
+            var existingPartType = await _storageProvider.GetPartTypeAsync(partType.PartTypeId, _requestContext.GetUserContext());
+            if (existingPartType == null)
+                return false;
+            
+            // does it have children? todo: need storage provider support for this
+            var partTypes = await GetPartTypesAsync();
+            if (partTypes.Where(x => x.ParentPartTypeId== existingPartType.PartTypeId).Any())
+            {
+                throw new InvalidOperationException($"Cannot delete part type '{existingPartType.Name}' until it's children are deleted.");
+            }            
+
             return await _storageProvider.DeletePartTypeAsync(partType, _requestContext.GetUserContext());
         }
